// src/lib/api/schema.d.ts
export interface components {
    schemas: {
// ...
        WorkoutRead: {
            /**
             * Id
             * Format: uuid
             */
            id: string;
            /**
             * Student Id
             * Format: uuid
             */
            student_id: string;
            /** Name */
            name: string;
            /** Instructions */
            instructions: string | null;
            /** Is Active */
            is_active: boolean;
            /** Exercises */
            exercises: components["schemas"]["WorkoutItemRead"][];
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}